+++ routes/tratamiento.py
@@ -2,6 +2,7 @@ from fastapi import APIRouter, HTTPException, status, Path
 from models.tratamiento import Tratamientos
 from schemas.tratamiento import TratamientoRequest
 from config.database import db_dependency
+from routes.token import user_dependecy
 
 
 router = APIRouter(
@@ -10,19 +11,29 @@ router = APIRouter(
 
 
 @router.get('/tratamientos', status_code=status.HTTP_200_OK)
-async def get_all_tratamientos(db: db_dependency):
+async def get_all_tratamientos(usuario: user_dependecy, db: db_dependency):
     
+    if usuario is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Autenticacion Fallida')
+
     return db.query(Tratamientos).all()
 
 
 @router.get('/tratamiento/{tratamiento_id}', status_code=status.HTTP_200_OK)
-async def get_tratamiento_by_id(db: db_dependency, tratamiento_id: int = Path(gt=0)):
+async def get_tratamiento_by_id(usuario: user_dependecy, db: db_dependency, tratamiento_id: int = Path(gt=0)):
     
+    if usuario is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Autenticacion Fallida')
+
     return db.query(Tratamientos).filter(Tratamientos.id == tratamiento_id).first()
 
 
 @router.post('/tratamientos', status_code=status.HTTP_201_CREATED)
-async def create_tratamiento(db: db_dependency, tratamiento_request: TratamientoRequest):
+async def create_tratamiento(usuario: user_dependecy, db: db_dependency, tratamiento_request: TratamientoRequest):
+
+    if usuario is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Autenticacion Fallida')
+
     tratamiento_model = Tratamientos(**tratamiento_request.model_dump())
 
     db.add(tratamiento_model)
@@ -30,7 +41,10 @@ async def create_tratamiento(db: db_dependency, tratamiento_request: Tratamiento
 
 
 @router.put('/tratamiento/{tratamiento_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def update_tratamiento(db: db_dependency, tratamiento_request: TratamientoRequest, tratamiento_id: int = Path(gt=0)):
+async def update_tratamiento(usuario: user_dependecy, db: db_dependency, tratamiento_request: TratamientoRequest, tratamiento_id: int = Path(gt=0)):
+
+    if usuario is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Autenticacion Fallida')
 
     tratamiento_model = db.query(Tratamientos).filter(Tratamientos.id == tratamiento_id).first()
 
@@ -48,7 +62,10 @@ async def update_tratamiento(db: db_dependency, tratamiento_request: Tratamiento
 
 
 @router.delete('/tratamiento/{tratamiento_id}', status_code=status.HTTP_204_NO_CONTENT)
-async def delete_tratamiento(db: db_dependency, tratamiento_id: int = Path(gt=0)):
+async def delete_tratamiento(usuario: user_dependecy, db: db_dependency, tratamiento_id: int = Path(gt=0)):
     
+    if usuario is None:
+        raise HTTPException(status_code=status.HTTP_401_UNAUTHORIZED, detail='Autenticacion Fallida')
+
     db.query(Tratamientos).filter(Tratamientos.id == tratamiento_id).delete()
     db.commit()
